fix:updated return type from 'T | null' to 'T'
