+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [4.0.0] - 2025-03-20
+
+### Changed
+
+- Updated result return type from `T | null` to `T`.
+
+## [3.0.1] - 2025-03-17
+
+### Added
+
+- Added CI/CD workflow.
+
 ## [3.0.0] - 2025-02-15
 
 ### Added
@@ -31,6 +43,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Introduced `x8tSync` function to handle synchronous functions with error handling.
+
+## [1.1.0] - 2024-10-30
 
 ### Fixed
