+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "x8t",
-  "version": "3.0.1",
+  "version": "3.1.0",
   "description": "A utility for safely executing functions.",
   "main": "dist/index.js",
   "author": "Marvin Ronquillo",
